Preserve context values
Instead of passing a fresh `context.Background()` wrap context passed by
the user with a context that ignores cancellation.

This allows to drop the original cancellation while preserving context
values.

If built with Go 1.21+ the stdlib `context.WithoutCancel` is used.
Otherwise it fallbacks to an in-tree copy of the withoutCancel.

Signed-off-by: Paweł Gronowski <pawel.gronowski@docker.com>